refactor(api): split expenses.ts and recurrence-series.ts into module directories
- Split monolithic expenses.ts (1720 lines) into expenses/ directory with
  dedicated modules: create, update, delete, queries, helpers, etc.
- Split recurrence-series.ts into recurrence/ directory with materialize,
  series-ops, template, and boss modules
- Add series-progress module for tracking recurring series generation
- Add tRPC procedure and integration tests for bulk recurring expenses
- Update exports and remaining references
