+++ apps/api/src/integration/recurring-bulk.test.ts
@@ -1,0 +1,572 @@
+import { GroupMemberStatus, GroupRole, prisma } from '@spliit/db'
+import { afterAll, afterEach, beforeAll, describe, expect, it } from 'vitest'
+import { randomId } from '../lib/api'
+import {
+  setDefaultActivityNotificationDispatchers,
+  waitForScheduledNotificationDispatchesForTest,
+} from '../lib/notifications/dispatcher'
+import { groupsRouter } from '../trpc/routers/groups'
+import { CapturingDispatcher, checkDbConnection, testRunId } from './setup'
+
+await checkDbConnection()
+
+/**
+ * THIS_AND_FUTURE update/delete — per-row activity + notification correctness.
+ *
+ * Covers review findings F1–F5:
+ *   1. Future-row update diffs must be derived from a full snapshot diff,
+ *      with real `changedFields`/`changes` and a per-row `affectedParticipants`
+ *      union (NOT the hard-coded `['recurrence']`).
+ *   2. When exactly one future row differs and the selected row already
+ *      matches the new template, the single-row notification must still
+ *      fire (sole-changed-future-row case).
+ *   3. Notes-only updates to future rows must log a `notes` field change,
+ *      not just `recurrence`.
+ *   4. THIS_AND_FUTURE delete snapshots must include documents on later
+ *      occurrences; cadence fields must surface in the summary data.
+ */
+describe('Recurring bulk updates — real DB', () => {
+  const runId = testRunId()
+  const adminId = `acct-rb-${runId}`
+  const adminEmail = `rb-${runId}@test.example`
+  const witnessId = `acct-rb-w-${runId}`
+  const witnessEmail = `rb-w-${runId}@test.example`
+
+  const ledgerIds: string[] = []
+  function trackLedger(id: string) {
+    ledgerIds.push(id)
+  }
+
+  function makeCaller() {
+    return groupsRouter.createCaller({
+      auth: {
+        session: { id: 'sess-test' },
+        user: {
+          id: adminId,
+          email: adminEmail,
+          emailVerified: true,
+          name: 'Test Admin',
+        },
+      },
+    } as never)
+  }
+
+  beforeAll(async () => {
+    await prisma.account.upsert({
+      where: { email: adminEmail },
+      update: {},
+      create: {
+        id: adminId,
+        email: adminEmail,
+        emailVerified: true,
+        name: 'Test Admin',
+      },
+    })
+    await prisma.account.upsert({
+      where: { email: witnessEmail },
+      update: {},
+      create: {
+        id: witnessId,
+        email: witnessEmail,
+        emailVerified: true,
+        name: 'Test Witness',
+      },
+    })
+  })
+
+  afterEach(async () => {
+    try {
+      await waitForScheduledNotificationDispatchesForTest()
+    } finally {
+      setDefaultActivityNotificationDispatchers([])
+    }
+  })
+
+  afterAll(async () => {
+    for (const lid of ledgerIds) {
+      await prisma.ledger.delete({ where: { id: lid } }).catch(() => {})
+    }
+    await prisma.account.delete({ where: { id: adminId } }).catch(() => {})
+    await prisma.account.delete({ where: { id: witnessId } }).catch(() => {})
+  })
+
+  /**
+   * Create a group with admin + extra UNLINKED participants. The admin
+   * comes from the tRPC caller; extras are inserted directly as
+   * UNLINKED_PARTICIPANT ledger rows (matches the multi-payer harness).
+   * Returns the group id, ledger id, and a name → ledgerParticipantId map.
+   */
+  async function createGroupWithParticipants(
+    name: string,
+    extraParticipantNames: string[],
+  ): Promise<{
+    groupId: string
+    ledgerId: string
+    participants: Record<string, string>
+  }> {
+    const caller = makeCaller()
+    const { groupId } = await caller.create({
+      groupFormValues: {
+        name,
+        currency: '$',
+        currencyCode: 'USD',
+        participants: [{ name: 'Admin' }],
+      },
+    })
+    const group = await prisma.group.findUnique({
+      where: { id: groupId },
+      include: {
+        ledger: true,
+        members: { include: { ledgerParticipant: true } },
+      },
+    })
+    trackLedger(group!.ledger.id)
+    const ledgerId = group!.ledger.id
+    const adminLpId = group!.members[0].ledgerParticipant!.id
+    const participants: Record<string, string> = { Admin: adminLpId }
+    for (const participantName of extraParticipantNames) {
+      const lp = await prisma.ledgerParticipant.create({
+        data: {
+          id: randomId(),
+          ledgerId,
+          kind: 'UNLINKED_PARTICIPANT',
+          displayName: participantName,
+        },
+      })
+      participants[participantName] = lp.id
+    }
+    return { groupId, ledgerId, participants }
+  }
+
+  /**
+   * Add a real account-based member (active) to a freshly-created
+   * group. Returns that member's ledgerParticipantId, so tests can
+   * exercise the notification path that excludes the actor — only
+   * account-backed LedgerParticipants resolve to a recipient.
+   */
+  async function addActiveMember(
+    groupId: string,
+    ledgerId: string,
+  ): Promise<string> {
+    const memberId = `gm-rb-w-${runId}-${randomId()}`
+    const gm = await prisma.groupMember.create({
+      data: {
+        id: memberId,
+        groupId,
+        accountId: witnessId,
+        role: GroupRole.MEMBER,
+        status: GroupMemberStatus.ACTIVE,
+        joinedAt: new Date(),
+      },
+    })
+    const lp = await prisma.ledgerParticipant.create({
+      data: {
+        id: `lp-rb-w-${runId}-${randomId()}`,
+        ledgerId,
+        groupMemberId: gm.id,
+      },
+    })
+    return lp.id
+  }
+
+  /**
+   * Materialize a recurring expense past its next occurrence boundary, so
+   * the recurring series has at least the original occurrence + 1 future
+   * occurrence persisted. Returns the ids in creation order.
+   */
+  async function seedSeriesWithFutureOccurrences(args: {
+    groupId: string
+    participants: Record<string, string>
+    title: string
+    initialPaidFor: Array<{ participant: string; shares: number }>
+  }): Promise<{
+    seriesId: string
+    expenseIds: string[]
+  }> {
+    const caller = makeCaller()
+    const pastDate = new Date()
+    pastDate.setUTCDate(pastDate.getUTCDate() - 14)
+    const result = await caller.expenses.create({
+      groupId: args.groupId,
+      expense: {
+        title: args.title,
+        amount: 6000,
+        expenseDate: pastDate.toISOString(),
+        category: 'general',
+        splitMode: 'EVENLY',
+        paidBySplitMode: 'BY_AMOUNT',
+        paidByList: [{ participant: args.participants.Admin, shares: 6000 }],
+        paidFor: args.initialPaidFor,
+        isReimbursement: false,
+        documents: [],
+        recurrenceRule: 'WEEKLY',
+      },
+    })
+    const originalId = result.expenseId
+    const { materializeRecurringExpense } =
+      await import('../lib/api/recurrence-series')
+    let series = await prisma.recurringExpenseSeries.findFirst({
+      where: {
+        ledgerId: (
+          await prisma.expense.findUniqueOrThrow({
+            where: { id: originalId },
+            select: { ledgerId: true },
+          })
+        ).ledgerId,
+        template: { path: ['title'], equals: args.title },
+      },
+    })
+    while (series && series.nextOccurrenceDate <= new Date()) {
+      await materializeRecurringExpense({
+        seriesId: series.id,
+        sequence: series.occurrencesCreated + 1,
+        occurrenceDate: series.nextOccurrenceDate.toISOString().slice(0, 10),
+      })
+      series = await prisma.recurringExpenseSeries.findUnique({
+        where: { id: series.id },
+      })
+    }
+    const cloned = await prisma.expense.findMany({
+      where: { ledgerId: series!.ledgerId, title: args.title },
+      orderBy: { createdAt: 'asc' },
+      select: { id: true, recurrenceSequence: true },
+    })
+    return { seriesId: series!.id, expenseIds: cloned.map((c) => c.id) }
+  }
+
+  // ------------------------------------------------------------------------
+  // F1 — Future-row paidFor change logs split/payers diff, not 'recurrence'
+  // ------------------------------------------------------------------------
+  it('THIS_AND_FUTURE paidFor change on future rows logs split (not recurrence)', async () => {
+    const { groupId, participants } = await createGroupWithParticipants(
+      `RB-F1-${runId}`,
+      ['Bob'],
+    )
+    const { expenseIds } = await seedSeriesWithFutureOccurrences({
+      groupId,
+      participants,
+      title: 'Pizza weekly',
+      initialPaidFor: [
+        { participant: participants.Admin, shares: 1 },
+        { participant: participants.Bob, shares: 1 },
+      ],
+    })
+    expect(expenseIds.length).toBeGreaterThanOrEqual(2)
+    const targetId = expenseIds[0]!
+    const futureIds = expenseIds.slice(1)
+
+    await makeCaller().expenses.update({
+      groupId,
+      expenseId: targetId,
+      scope: 'THIS_AND_FUTURE',
+      expense: {
+        title: 'Pizza weekly',
+        amount: 6000,
+        expenseDate: (
+          await prisma.expense.findUniqueOrThrow({
+            where: { id: targetId },
+            select: { expenseDate: true },
+          })
+        ).expenseDate.toISOString(),
+        category: 'general',
+        splitMode: 'EVENLY',
+        paidBySplitMode: 'BY_AMOUNT',
+        paidByList: [{ participant: participants.Admin, shares: 6000 }],
+        // Drop Bob from paidFor on the new template.
+        paidFor: [{ participant: participants.Admin, shares: 1 }],
+        isReimbursement: false,
+        documents: [],
+        recurrenceRule: 'WEEKLY',
+        recurrence: {
+          frequency: 'WEEKLY',
+          interval: 1,
+          end: { type: 'INDEFINITE' },
+        },
+      },
+    })
+
+    // The future rows must each carry an EXPENSE_UPDATED activity whose
+    // changedFields flags 'split' or 'payers' (never 'recurrence'), with
+    // affectedParticipants containing both Admin (still present) and Bob
+    // (removed — but recipients still cover removed-by-change members).
+    for (const futureId of futureIds) {
+      const activity = await prisma.activity.findFirst({
+        where: { subjectId: futureId, type: 'EXPENSE_UPDATED' },
+        orderBy: { time: 'desc' },
+      })
+      expect(activity).not.toBeNull()
+      const data = activity!.data as Record<string, unknown>
+      const changedFields = data.changedFields as string[] | undefined
+      expect(changedFields ?? []).not.toContain('recurrence')
+      // The split differ flags the change either via 'split' or 'payers'.
+      expect(changedFields ?? []).toEqual(expect.arrayContaining(['split']))
+      const participantsList = data.affectedParticipants as string[] | undefined
+      expect(participantsList).toEqual(
+        expect.arrayContaining([participants.Admin, participants.Bob]),
+      )
+    }
+  })
+
+  // ------------------------------------------------------------------------
+  // F2 — Sole changed future row produces a single notification payload
+  // ------------------------------------------------------------------------
+  it('THIS_AND_FUTURE where only a future row differs logs exactly one activity', async () => {
+    const { groupId, participants } = await createGroupWithParticipants(
+      `RB-F2-${runId}`,
+      ['Bob'],
+    )
+    const { expenseIds } = await seedSeriesWithFutureOccurrences({
+      groupId,
+      participants,
+      title: 'Streaming',
+      initialPaidFor: [
+        { participant: participants.Admin, shares: 1 },
+        { participant: participants.Bob, shares: 1 },
+      ],
+    })
+    expect(expenseIds.length).toBeGreaterThanOrEqual(3)
+    const targetId = expenseIds[0]!
+    const laterId = expenseIds[2]!
+
+    // Manually adjust a single later row to pre-divert it from the new
+    // template: Bob dropped, title left intact. After the THIS_AND_FUTURE
+    // update the selected row matches the new template exactly, and only
+    // the manually-adjusted later row will see a real diff.
+    await prisma.expense.update({
+      where: { id: laterId },
+      data: { title: 'Old streaming' },
+    })
+
+    const beforeSelectedActivities = await prisma.activity.count({
+      where: { subjectId: targetId, type: 'EXPENSE_UPDATED' },
+    })
+
+    await makeCaller().expenses.update({
+      groupId,
+      expenseId: targetId,
+      scope: 'THIS_AND_FUTURE',
+      expense: {
+        // Reverts the manual title edit on the later row.
+        title: 'Streaming',
+        amount: 6000,
+        expenseDate: (
+          await prisma.expense.findUniqueOrThrow({
+            where: { id: targetId },
+            select: { expenseDate: true },
+          })
+        ).expenseDate.toISOString(),
+        category: 'general',
+        splitMode: 'EVENLY',
+        paidBySplitMode: 'BY_AMOUNT',
+        paidByList: [{ participant: participants.Admin, shares: 6000 }],
+        paidFor: [
+          { participant: participants.Admin, shares: 1 },
+          { participant: participants.Bob, shares: 1 },
+        ],
+        isReimbursement: false,
+        documents: [],
+        recurrenceRule: 'WEEKLY',
+        recurrence: {
+          frequency: 'WEEKLY',
+          interval: 1,
+          end: { type: 'INDEFINITE' },
+        },
+      },
+    })
+
+    // The selected expense itself matches the new template — no new
+    // EXPENSE_UPDATED activity for it.
+    const afterSelectedActivities = await prisma.activity.count({
+      where: { subjectId: targetId, type: 'EXPENSE_UPDATED' },
+    })
+    expect(afterSelectedActivities).toBe(beforeSelectedActivities)
+
+    // Exactly one future-row activity — the manually-adjusted row's title
+    // being reverted to the new template.
+    const futureRowActivities = await prisma.activity.findMany({
+      where: {
+        type: 'EXPENSE_UPDATED',
+        subjectId: { in: expenseIds.slice(1) },
+      },
+    })
+    expect(futureRowActivities).toHaveLength(1)
+    const soleActivity = futureRowActivities[0]!
+    expect(soleActivity.subjectId).toBe(laterId)
+    const soleData = soleActivity.data as Record<string, unknown>
+    expect(soleData.title).toBe('Streaming')
+    expect((soleData.changedFields as string[]) ?? []).toEqual(
+      expect.arrayContaining(['title']),
+    )
+  })
+
+  // ------------------------------------------------------------------------
+  // F3 — Notes-only update on future rows logs 'notes' (not 'recurrence')
+  // ------------------------------------------------------------------------
+  it('THIS_AND_FUTURE notes-only change logs notes on future rows', async () => {
+    const { groupId, participants } = await createGroupWithParticipants(
+      `RB-F3-${runId}`,
+      [],
+    )
+    const { expenseIds } = await seedSeriesWithFutureOccurrences({
+      groupId,
+      participants,
+      title: 'Notes test',
+      initialPaidFor: [{ participant: participants.Admin, shares: 1 }],
+    })
+    expect(expenseIds.length).toBeGreaterThanOrEqual(2)
+    const targetId = expenseIds[0]!
+    const futureIds = expenseIds.slice(1)
+
+    await makeCaller().expenses.update({
+      groupId,
+      expenseId: targetId,
+      scope: 'THIS_AND_FUTURE',
+      expense: {
+        title: 'Notes test',
+        amount: 6000,
+        expenseDate: (
+          await prisma.expense.findUniqueOrThrow({
+            where: { id: targetId },
+            select: { expenseDate: true },
+          })
+        ).expenseDate.toISOString(),
+        category: 'general',
+        splitMode: 'EVENLY',
+        paidBySplitMode: 'BY_AMOUNT',
+        paidByList: [{ participant: participants.Admin, shares: 6000 }],
+        paidFor: [{ participant: participants.Admin, shares: 1 }],
+        isReimbursement: false,
+        notes: 'added by test',
+        documents: [],
+        recurrenceRule: 'WEEKLY',
+        recurrence: {
+          frequency: 'WEEKLY',
+          interval: 1,
+          end: { type: 'INDEFINITE' },
+        },
+      },
+    })
+
+    for (const futureId of futureIds) {
+      const activity = await prisma.activity.findFirst({
+        where: { subjectId: futureId, type: 'EXPENSE_UPDATED' },
+        orderBy: { time: 'desc' },
+      })
+      expect(activity).not.toBeNull()
+      const data = activity!.data as Record<string, unknown>
+      expect((data.changedFields as string[]) ?? []).toEqual(
+        expect.arrayContaining(['notes']),
+      )
+      expect((data.changedFields as string[]) ?? []).not.toContain('recurrence')
+    }
+  })
+
+  // ------------------------------------------------------------------------
+  // F3b — Reimbursement-only update on future rows logs 'reimbursement'
+  // ------------------------------------------------------------------------
+  it('THIS_AND_FUTURE reimbursement-only change logs reimbursement on future rows', async () => {
+    const { groupId, participants } = await createGroupWithParticipants(
+      `RB-F3b-${runId}`,
+      [],
+    )
+    const { expenseIds } = await seedSeriesWithFutureOccurrences({
+      groupId,
+      participants,
+      title: 'Reimbursement test',
+      initialPaidFor: [{ participant: participants.Admin, shares: 1 }],
+    })
+    expect(expenseIds.length).toBeGreaterThanOrEqual(2)
+    const targetId = expenseIds[0]!
+    const futureIds = expenseIds.slice(1)
+
+    await makeCaller().expenses.update({
+      groupId,
+      expenseId: targetId,
+      scope: 'THIS_AND_FUTURE',
+      expense: {
+        title: 'Reimbursement test',
+        amount: 6000,
+        expenseDate: (
+          await prisma.expense.findUniqueOrThrow({
+            where: { id: targetId },
+            select: { expenseDate: true },
+          })
+        ).expenseDate.toISOString(),
+        category: 'general',
+        splitMode: 'EVENLY',
+        paidBySplitMode: 'BY_AMOUNT',
+        paidByList: [{ participant: participants.Admin, shares: 6000 }],
+        paidFor: [{ participant: participants.Admin, shares: 1 }],
+        isReimbursement: true,
+        documents: [],
+        recurrenceRule: 'WEEKLY',
+        recurrence: {
+          frequency: 'WEEKLY',
+          interval: 1,
+          end: { type: 'INDEFINITE' },
+        },
+      },
+    })
+
+    for (const futureId of futureIds) {
+      const activity = await prisma.activity.findFirst({
+        where: { subjectId: futureId, type: 'EXPENSE_UPDATED' },
+        orderBy: { time: 'desc' },
+      })
+      expect(activity).not.toBeNull()
+      const data = activity!.data as Record<string, unknown>
+      expect((data.changedFields as string[]) ?? []).toEqual(
+        expect.arrayContaining(['reimbursement']),
+      )
+      expect((data.changedFields as string[]) ?? []).not.toContain('recurrence')
+    }
+  })
+
+  // ------------------------------------------------------------------------
+  // F5 — THIS_AND_FUTURE delete summary exposes cadence fields
+  // ------------------------------------------------------------------------
+  it('THIS_AND_FUTURE delete summary carries frequency/interval/endType/seriesEndDate', async () => {
+    // Two paidFor participants so the summary notification reaches
+    // somebody other than the admin actor (handlers exclude the actor).
+    const { groupId, ledgerId, participants } =
+      await createGroupWithParticipants(`RB-F5-${runId}`, [])
+    const witnessLpId = await addActiveMember(groupId, ledgerId)
+    const witnesses = { ...participants, Witness: witnessLpId }
+    const { expenseIds } = await seedSeriesWithFutureOccurrences({
+      groupId,
+      participants: witnesses,
+      title: 'Rent weekly',
+      initialPaidFor: [
+        { participant: witnesses.Admin, shares: 1 },
+        { participant: witnesses.Witness, shares: 1 },
+      ],
+    })
+    expect(expenseIds.length).toBeGreaterThanOrEqual(2)
+    const targetId = expenseIds[0]!
+
+    const capture = new CapturingDispatcher()
+    setDefaultActivityNotificationDispatchers([capture])
+
+    await makeCaller().expenses.delete({
+      groupId,
+      expenseId: targetId,
+      scope: 'THIS_AND_FUTURE',
+    })
+
+    await waitForScheduledNotificationDispatchesForTest()
+    const summaryEvent = capture.events.find(
+      (e) => e.type === 'EXPENSE_DELETED' && e.subject === null,
+    )
+    expect(summaryEvent).toBeDefined()
+    const data = summaryEvent!.data as Record<string, unknown>
+    expect(data.kind).toBe('recurring_expense_summary')
+    expect(data.frequency).toBe('WEEKLY')
+    expect(data.interval).toBe(1)
+    expect(data.endType).toBe('INDEFINITE')
+    expect(data.operation).toBe('delete')
+    expect(typeof data.count).toBe('number')
+    expect((data.count as number) >= 2).toBe(true)
+    expect(typeof data.startDate).toBe('string')
+    expect(typeof data.endDate).toBe('string')
+  })
+})
