--- apps/api/src/lib/api/expenses.ts
@@ -1,1720 +1,0 @@
-import type { Prisma } from '@spliit/db'
-import { prisma, type Expense as DbExpense } from '@spliit/db'
-import {
-  calculateRecurrenceDate,
-  categoryIdSchema,
-  commonCurrencyLookbackDate,
-  computePaidForFromItems,
-  conversionFromStored,
-  DEFAULT_CATEGORY_ID,
-  getCategoryById,
-  rankCommonCurrencies,
-  type Category,
-  type CategoryId,
-  type Expense,
-} from '@spliit/domain'
-import { env as jobsEnv } from '@spliit/jobs'
-import { deleteS3Object, promoteUploadedDocument } from '../../routes/upload'
-import { resolveConversion } from '../expense-conversion'
-import { resolveParticipantDisplayName } from '../invitations'
-import { scheduleDefaultNotificationDispatch } from '../notifications/dispatcher'
-import { buildExpenseActivityData, logActivity } from './activities'
-import {
-  getAffectedParticipantIds,
-  getExpenseChangeSummary,
-  type ChangeContext,
-} from './expense-activity-diff'
-import {
-  buildRecurringTemplate,
-  createSeriesForExpense,
-  enqueueMaterialization,
-  getApiBoss,
-  getExpenseRecurrence,
-  toRecurrenceConfig,
-} from './recurrence-series'
-import { randomId } from './shared'
-
-/**
- * Resolve a `categoryId` string from the database to the in-code
- * {@link Category} object. Returns the default "General" category when
- * the stored id is not in the in-code list (e.g. it was written by an
- * older version of the app or is otherwise invalid).
- */
-function resolveCategory(categoryId: string): Category {
-  const parsedCategoryId = categoryIdSchema.safeParse(categoryId)
-  return (
-    (parsedCategoryId.success
-      ? getCategoryById(parsedCategoryId.data)
-      : undefined) ?? getCategoryById(DEFAULT_CATEGORY_ID)!
-  )
-}
-
-/**
- * Narrow a `categoryId` string from the database to the {@link CategoryId}
- * literal union, falling back to the default category if the stored id is
- * not in the in-code list.
- */
-function narrowCategoryId(categoryId: string): CategoryId {
-  const parsed = categoryIdSchema.safeParse(categoryId)
-  return parsed.success ? parsed.data : DEFAULT_CATEGORY_ID
-}
-
-/**
- * Normalize the Prisma `getExpense` return value to the domain
- * `Expense` shape expected by diff and affected-participant utilities.
- * The Prisma model stores `ledgerParticipantId` while the domain uses
- * `participant` for payer / split / item references.
- */
-/**
- * Map a stored expense into the shape used by activity diffs.
- * `amount` is the ledger total; flat conversion fields are attached for
- * amount/conversion differs (in addition to the `conversion` discriminant).
- */
-function toExpenseDomainShape(
-  existing: NonNullable<Awaited<ReturnType<typeof getExpense>>>,
-): Expense & {
-  originalAmount?: number
-  originalCurrency?: string
-  conversionRate?: number
-  conversionSource?: 'EXCHANGE' | 'CUSTOM' | null
-} {
-  return {
-    title: existing.title,
-    amount: existing.amount,
-    expenseDate: existing.expenseDate,
-    category: existing.categoryId as Expense['category'],
-    notes: existing.notes ?? undefined,
-    recurrenceRule: existing.recurringSeries?.frequency ?? 'NONE',
-    recurrence: existing.recurringSeries
-      ? toRecurrenceConfig(existing.recurringSeries)
-      : null,
-    splitMode: existing.splitMode,
-    paidBySplitMode: existing.paidBySplitMode,
-    paidByList: existing.paidByList.map((pb) => ({
-      participant: pb.ledgerParticipantId,
-      shares: pb.shares,
-    })),
-    paidFor: existing.paidFor.map((pf) => ({
-      participant: pf.ledgerParticipantId,
-      shares: pf.shares,
-    })),
-    items: (existing.items ?? []).map((item) => ({
-      id: item.id,
-      title: item.title,
-      unitPrice: item.unitPrice,
-      quantity: item.quantity,
-      amount: item.amount,
-      splitMode: item.splitMode,
-      paidFor: item.paidFor.map((pf) => ({
-        participant: pf.ledgerParticipantId,
-        shares: pf.shares,
-      })),
-    })),
-    itemizedRemainder: existing.itemizedRemainder
-      ? {
-          splitMode: existing.itemizedRemainder.splitMode,
-          paidFor: existing.itemizedRemainder.paidFor.map((pf) => ({
-            participant: pf.ledgerParticipantId,
-            shares: pf.shares,
-          })),
-        }
-      : undefined,
-    documents: existing.documents.map((d) => ({
-      id: d.id,
-      url: d.url,
-      width: d.width,
-      height: d.height,
-    })),
-    conversion: conversionFromStored({
-      conversionSource: existing.conversionSource,
-      originalCurrency: existing.originalCurrency,
-      conversionRate: existing.conversionRate,
-    }),
-    originalAmount: existing.originalAmount ?? undefined,
-    originalCurrency: existing.originalCurrency ?? undefined,
-    conversionRate: existing.conversionRate ?? undefined,
-    conversionSource: existing.conversionSource,
-    isReimbursement: existing.isReimbursement,
-  } as Expense & {
-    originalAmount?: number
-    originalCurrency?: string
-    conversionRate?: number
-    conversionSource?: 'EXCHANGE' | 'CUSTOM' | null
-  }
-}
-
-async function promoteExpenseDocuments(
-  documents: Array<{ id: string; url: string; width: number; height: number }>,
-): Promise<Array<{ id: string; url: string; width: number; height: number }>> {
-  return Promise.all(
-    documents.map(async (doc) => ({
-      ...doc,
-      url: await promoteUploadedDocument(doc.url),
-    })),
-  )
-}
-
-export async function createExpense(
-  expense: Expense,
-  groupId: string,
-  actor: { accountId: string },
-): Promise<DbExpense> {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    include: { ledger: true },
-  })
-  if (!group || !group.ledgerId) throw new Error(`Invalid group ID: ${groupId}`)
-
-  const ledgerId = group.ledgerId
-
-  const conversion = await resolveConversion(expense, {
-    ledgerCurrency: group.ledger.currencyCode ?? null,
-    expenseDate: expense.expenseDate,
-  })
-
-  const expenseAmount = conversion.ledgerAmountMinor
-
-  const participants = await prisma.ledgerParticipant.findMany({
-    where: {
-      ledgerId,
-      OR: [
-        { groupMemberId: { not: null } },
-        { invitations: { some: { status: 'PENDING' } } },
-        { kind: 'UNLINKED_PARTICIPANT' },
-      ],
-    },
-    select: { id: true },
-  })
-  const participantIds = new Set(participants.map((p) => p.id))
-
-  for (const participantId of [
-    ...expense.paidByList.map((p) => p.participant),
-    ...expense.paidFor.map((p) => p.participant),
-    ...(expense.items ?? []).flatMap((item) =>
-      item.paidFor.map((p) => p.participant),
-    ),
-    ...(expense.itemizedRemainder?.paidFor ?? []).map((p) => p.participant),
-  ]) {
-    if (!participantIds.has(participantId)) {
-      throw new Error(`Invalid participant ID: ${participantId}`)
-    }
-  }
-
-  const expenseId = randomId()
-
-  const expenseDateStr = expense.expenseDate.toISOString().slice(0, 10)
-
-  const recurrence = getExpenseRecurrence(
-    expense as unknown as { recurrence?: unknown; recurrenceRule?: string },
-    expense.expenseDate,
-  )
-  const isCreateRecurrence = recurrence !== null
-  const queueBoss =
-    recurrence && jobsEnv.JOBS_ENABLED ? await getApiBoss() : undefined
-
-  const documents = await promoteExpenseDocuments(expense.documents)
-  const recurringPaidFor =
-    expense.splitMode === 'ITEMIZED'
-      ? computePaidForFromItems(
-          expense.items ?? [],
-          [...participantIds],
-          conversion.originalAmount ?? expenseAmount,
-          expense.itemizedRemainder,
-          expenseId,
-        ).paidFor
-      : expense.paidFor
-
-  const { activity, createdExpense } = await prisma.$transaction(async (tx) => {
-    await tx.$queryRaw`SELECT id FROM "Group" WHERE id = ${groupId} FOR UPDATE`
-    const lockedGroup = await tx.group.findUnique({
-      where: { id: groupId },
-      select: { archived: true },
-    })
-    if (lockedGroup?.archived)
-      throw new Error('This group is archived and no new expenses can be added')
-    const activity = await logActivity(
-      groupId,
-      {
-        type: 'EXPENSE_CREATED',
-        actor: { type: 'ACCOUNT', id: actor.accountId },
-        subject: { type: 'EXPENSE', id: expenseId },
-        data: buildExpenseActivityData({
-          summary: expense.title,
-          title: expense.title,
-          amount: expenseAmount,
-          currencyCode: conversion.originalCurrency,
-          date: expenseDateStr,
-          originalAmount: conversion.originalAmount ?? undefined,
-          conversionRate: conversion.conversionRate ?? undefined,
-          conversionSource: conversion.conversionSource,
-          ledgerCurrencyCode: group.ledger.currencyCode ?? null,
-        }),
-      },
-      tx,
-    )
-
-    const recurringSeriesId = isCreateRecurrence ? randomId() : undefined
-    if (recurrence && recurringSeriesId) {
-      await createSeriesForExpense({
-        tx,
-        seriesId: recurringSeriesId,
-        ledgerId,
-        creatorAccountId: actor.accountId,
-        anchorDate: expense.expenseDate,
-        config: recurrence,
-        template: buildRecurringTemplate({
-          expense: { ...expense, paidForOverride: recurringPaidFor },
-          conversion,
-        }),
-        boss: queueBoss,
-      })
-    }
-
-    const createdExpense = await tx.expense.create({
-      data: {
-        id: expenseId,
-        ledgerId,
-        expenseDate: expense.expenseDate,
-        categoryId: expense.category,
-        amount: expenseAmount,
-        originalAmount: conversion.originalAmount,
-        originalCurrency: conversion.originalCurrency,
-        conversionRate: conversion.conversionRate,
-        conversionSource: conversion.conversionSource,
-        title: expense.title,
-        paidBySplitMode: expense.paidBySplitMode,
-        paidByList: {
-          createMany: {
-            data: expense.paidByList.map((paidBy) => ({
-              ledgerParticipantId: paidBy.participant,
-              shares: paidBy.shares,
-            })),
-          },
-        },
-        splitMode: expense.splitMode,
-        ...(recurringSeriesId
-          ? { recurringSeriesId, recurrenceSequence: 1 }
-          : {}),
-        paidFor: {
-          createMany: {
-            data:
-              expense.splitMode === 'ITEMIZED'
-                ? computePaidForFromItems(
-                    expense.items ?? [],
-                    [...participantIds],
-                    conversion.originalAmount ?? expenseAmount,
-                    expense.itemizedRemainder,
-                    expenseId,
-                  ).paidFor.map((p) => ({
-                    ledgerParticipantId: p.participant,
-                    shares: p.shares,
-                  }))
-                : expense.paidFor.map((paidFor) => ({
-                    ledgerParticipantId: paidFor.participant,
-                    shares: paidFor.shares,
-                  })),
-          },
-        },
-        items: {
-          create: (expense.items ?? []).map((item) => ({
-            // Item IDs are database-global and create requests may be copied
-            // or replayed, so never persist a client-provided ID here.
-            id: randomId(),
-            title: item.title,
-            unitPrice: item.unitPrice,
-            quantity: item.quantity,
-            amount: item.amount,
-            splitMode: item.splitMode,
-            paidFor: {
-              createMany: {
-                data: item.paidFor.map((pf) => ({
-                  ledgerParticipantId: pf.participant,
-                  shares: pf.shares,
-                })),
-              },
-            },
-          })),
-        },
-        ...(expense.itemizedRemainder
-          ? {
-              itemizedRemainder: {
-                create: {
-                  splitMode: expense.itemizedRemainder.splitMode,
-                  paidFor: {
-                    createMany: {
-                      data: expense.itemizedRemainder.paidFor.map((pf) => ({
-                        ledgerParticipantId: pf.participant,
-                        shares: pf.shares,
-                      })),
-                    },
-                  },
-                },
-              },
-            }
-          : {}),
-        isReimbursement: expense.isReimbursement,
-        documents: {
-          createMany: {
-            data: documents.map((doc) => ({
-              id: randomId(),
-              url: doc.url,
-              width: doc.width,
-              height: doc.height,
-              ledgerId,
-            })),
-          },
-        },
-        notes: expense.notes,
-      },
-    })
-
-    return { activity, createdExpense }
-  })
-
-  scheduleDefaultNotificationDispatch({
-    activityId: activity.id,
-    type: 'EXPENSE_CREATED',
-    groupId,
-    actor: { type: 'ACCOUNT', id: actor.accountId },
-    subject: { type: 'EXPENSE', id: expenseId },
-    data: buildExpenseActivityData({
-      summary: expense.title,
-      title: expense.title,
-      amount: expenseAmount,
-      currencyCode: conversion.originalCurrency,
-      date: expenseDateStr,
-      originalAmount: conversion.originalAmount ?? undefined,
-      conversionRate: conversion.conversionRate ?? undefined,
-      conversionSource: conversion.conversionSource,
-      ledgerCurrencyCode: group.ledger.currencyCode ?? null,
-    }),
-    occurredAt: activity.time,
-  })
-
-  return createdExpense
-}
-
-export async function deleteExpense(
-  groupId: string,
-  expenseId: string,
-  actor: { accountId: string },
-  options?: {
-    scope?: 'OCCURRENCE' | 'THIS_AND_FUTURE'
-    /** For THIS_AND_FUTURE, also cancel the series after deleting rows. */
-    stopRecurrence?: boolean
-  },
-) {
-  const existingExpense = await getExpense(groupId, expenseId)
-  if (!existingExpense) throw new Error(`Invalid expense ID: ${expenseId}`)
-
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: { ledger: { select: { currencyCode: true } } },
-  })
-
-  const affectedParticipantIds = [
-    ...getAffectedParticipantIds({
-      oldExpense: toExpenseDomainShape(existingExpense),
-    }),
-  ]
-
-  const expenseDateStr = existingExpense.expenseDate.toISOString().slice(0, 10)
-
-  if (
-    options?.stopRecurrence !== undefined &&
-    options.scope !== 'THIS_AND_FUTURE'
-  ) {
-    throw new Error(
-      'stopRecurrence is only valid with THIS_AND_FUTURE deletion scope',
-    )
-  }
-  if (
-    options?.scope === 'THIS_AND_FUTURE' &&
-    !existingExpense.recurringSeriesId
-  ) {
-    throw new Error('THIS_AND_FUTURE deletion requires a recurring expense')
-  }
-
-  const activity = await prisma.$transaction(async (tx) => {
-    if (existingExpense.recurringSeriesId) {
-      await tx.$queryRaw`SELECT id FROM "RecurringExpenseSeries" WHERE id = ${existingExpense.recurringSeriesId} FOR UPDATE`
-    }
-    const act = await logActivity(
-      groupId,
-      {
-        type: 'EXPENSE_DELETED',
-        actor: { type: 'ACCOUNT', id: actor.accountId },
-        subject: { type: 'EXPENSE', id: expenseId },
-        data: buildExpenseActivityData({
-          summary: existingExpense.title,
-          title: existingExpense.title,
-          amount: existingExpense.amount,
-          currencyCode: existingExpense.originalCurrency ?? null,
-          date: expenseDateStr,
-          affectedParticipants: affectedParticipantIds,
-          originalAmount: existingExpense.originalAmount ?? undefined,
-          conversionRate: existingExpense.conversionRate
-            ? Number(existingExpense.conversionRate)
-            : undefined,
-          conversionSource: existingExpense.conversionSource,
-          ledgerCurrencyCode: group?.ledger.currencyCode ?? null,
-        }),
-      },
-      tx,
-    )
-
-    if (
-      options?.scope === 'THIS_AND_FUTURE' &&
-      existingExpense.recurringSeriesId &&
-      existingExpense.recurrenceSequence
-    ) {
-      await tx.expense.deleteMany({
-        where: {
-          ledgerId: existingExpense.ledgerId,
-          recurringSeriesId: existingExpense.recurringSeriesId,
-          recurrenceSequence: { gte: existingExpense.recurrenceSequence },
-        },
-      })
-      // Bump the optimistic-concurrency version even for delete-only. A
-      // materializer that took a snapshot before this mutation must not be
-      // allowed to recreate one of the rows just removed.
-      await tx.recurringExpenseSeries.update({
-        where: { id: existingExpense.recurringSeriesId },
-        data: {
-          version: { increment: 1 },
-          catchUpBatch: null,
-          ...(options.stopRecurrence ? { status: 'CANCELLED' as const } : {}),
-        },
-      })
-    } else {
-      await tx.expense.deleteMany({
-        where: { id: expenseId, ledgerId: existingExpense.ledgerId },
-      })
-    }
-
-    return act
-  })
-
-  // Best-effort S3 cleanup — errors are logged but not propagated.
-  for (const doc of existingExpense.documents) {
-    try {
-      await deleteS3Object(doc.url)
-    } catch (err) {
-      console.warn(`[expenses] failed to delete S3 object ${doc.url}:`, err)
-    }
-  }
-
-  scheduleDefaultNotificationDispatch({
-    activityId: activity.id,
-    type: 'EXPENSE_DELETED',
-    groupId,
-    actor: { type: 'ACCOUNT', id: actor.accountId },
-    subject: { type: 'EXPENSE', id: expenseId },
-    data: buildExpenseActivityData({
-      summary: existingExpense.title,
-      title: existingExpense.title,
-      amount: existingExpense.amount,
-      currencyCode: existingExpense.originalCurrency ?? null,
-      date: expenseDateStr,
-      affectedParticipants: affectedParticipantIds,
-      originalAmount: existingExpense.originalAmount ?? undefined,
-      conversionRate: existingExpense.conversionRate
-        ? Number(existingExpense.conversionRate)
-        : undefined,
-      conversionSource: existingExpense.conversionSource,
-      ledgerCurrencyCode: group?.ledger.currencyCode ?? null,
-    }),
-    occurredAt: activity.time,
-  })
-}
-
-/**
- * Stop a recurring series without deleting any already-materialized expenses.
- * The series row is the concurrency boundary; queued jobs will observe the
- * cancelled status and reconciliation will no longer enqueue work.
- */
-export async function stopRecurrence(
-  groupId: string,
-  expenseId: string,
-): Promise<void> {
-  const existingExpense = await getExpense(groupId, expenseId)
-  if (!existingExpense) throw new Error(`Invalid expense ID: ${expenseId}`)
-  if (!existingExpense.recurringSeriesId) {
-    throw new Error('Expense is not part of a recurring series')
-  }
-  const seriesId = existingExpense.recurringSeriesId
-
-  await prisma.$transaction(async (tx) => {
-    await tx.$queryRaw`SELECT id FROM "RecurringExpenseSeries" WHERE id = ${seriesId} FOR UPDATE`
-    const series = await tx.recurringExpenseSeries.findUnique({
-      where: { id: seriesId },
-      select: { status: true },
-    })
-    if (!series) throw new Error('Recurring series not found')
-    if (series.status === 'CANCELLED' || series.status === 'COMPLETED') return
-    await tx.recurringExpenseSeries.update({
-      where: { id: seriesId },
-      data: {
-        status: 'CANCELLED',
-        version: { increment: 1 },
-        catchUpBatch: null,
-      },
-    })
-  })
-}
-
-export async function updateExpense(
-  groupId: string,
-  expenseId: string,
-  expense: Expense,
-  actor: { accountId: string },
-  options?: { scope?: 'OCCURRENCE' | 'THIS_AND_FUTURE' },
-) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    include: { ledger: true },
-  })
-  if (!group || !group.ledgerId) throw new Error(`Invalid group ID: ${groupId}`)
-
-  const existingExpense = await getExpense(groupId, expenseId)
-  if (!existingExpense) throw new Error(`Invalid expense ID: ${expenseId}`)
-
-  const preserveRecurringDates =
-    options?.scope === 'THIS_AND_FUTURE' &&
-    existingExpense.recurringSeriesId !== null &&
-    existingExpense.recurrenceSequence !== null
-
-  const conversion = await resolveConversion(expense, {
-    ledgerCurrency: group.ledger.currencyCode ?? null,
-    expenseDate: preserveRecurringDates
-      ? existingExpense.expenseDate
-      : expense.expenseDate,
-  })
-
-  const expenseAmount = conversion.ledgerAmountMinor
-
-  const participants = await prisma.ledgerParticipant.findMany({
-    where: {
-      ledgerId: group.ledgerId,
-      OR: [
-        { groupMemberId: { not: null } },
-        { invitations: { some: { status: 'PENDING' } } },
-        { kind: 'UNLINKED_PARTICIPANT' },
-      ],
-    },
-    select: {
-      id: true,
-      displayName: true,
-      groupMember: {
-        select: {
-          account: { select: { id: true, name: true, image: true } },
-        },
-      },
-      invitations: {
-        where: { status: 'PENDING' },
-        select: { email: true, temporaryName: true },
-        take: 1,
-      },
-    },
-  })
-  const participantIds = new Set(participants.map((p) => p.id))
-  for (const participantId of [
-    ...expense.paidByList.map((p) => p.participant),
-    ...expense.paidFor.map((p) => p.participant),
-    ...(expense.items ?? []).flatMap((item) =>
-      item.paidFor.map((p) => p.participant),
-    ),
-    ...(expense.itemizedRemainder?.paidFor ?? []).map((p) => p.participant),
-  ]) {
-    if (!participantIds.has(participantId)) {
-      throw new Error(`Invalid participant ID: ${participantId}`)
-    }
-  }
-
-  // Build participant name map for change formatting.
-  const participantNameMap = new Map<string, string>()
-  for (const p of participants) {
-    participantNameMap.set(p.id, resolveParticipantDisplayName(p))
-  }
-
-  const changeCtx: ChangeContext = {
-    getParticipantName: (id: string) => participantNameMap.get(id) ?? id,
-    getCategoryName: (id: string) => resolveCategory(id).name,
-    formatCurrencyCents: (cents, currency) => {
-      const code =
-        currency ?? group.ledger.currencyCode ?? group.ledger.currency
-      const whole = Math.floor(Math.abs(cents) / 100)
-      const frac = Math.abs(cents) % 100
-      const sign = cents < 0 ? '-' : ''
-      return `${sign}${code} ${whole}.${frac.toString().padStart(2, '0')}`
-    },
-    ledgerCurrencyCode: group.ledger.currencyCode ?? group.ledger.currency,
-  }
-
-  const expenseDateStr = expense.expenseDate.toISOString().slice(0, 10)
-
-  const documents = await promoteExpenseDocuments(expense.documents)
-  const recurrence = getExpenseRecurrence(
-    expense as unknown as { recurrence?: unknown; recurrenceRule?: string },
-    expense.expenseDate,
-  )
-  const legacyRecurrenceRule =
-    recurrence?.frequency === 'DAILY' ||
-    recurrence?.frequency === 'WEEKLY' ||
-    recurrence?.frequency === 'MONTHLY'
-      ? recurrence.frequency
-      : ('NONE' as const)
-  const queueBoss =
-    recurrence && jobsEnv.JOBS_ENABLED ? await getApiBoss() : undefined
-
-  // Diff against the post-resolution shape so ledger `amount` and
-  // conversion fields are compared in the same units as the stored row.
-  const resolvedExpense = {
-    ...expense,
-    recurrenceRule: legacyRecurrenceRule,
-    amount: expenseAmount,
-    conversion: conversion.conversionSource
-      ? conversion.conversionSource === 'CUSTOM'
-        ? ({
-            type: 'custom' as const,
-            currency: conversion.originalCurrency ?? '',
-            rate: conversion.conversionRate ?? 1,
-          } as const)
-        : ({
-            type: 'exchange' as const,
-            currency: conversion.originalCurrency ?? '',
-          } as const)
-      : undefined,
-    originalAmount: conversion.originalAmount ?? undefined,
-    originalCurrency: conversion.originalCurrency ?? undefined,
-    conversionRate: conversion.conversionRate ?? undefined,
-    conversionSource: conversion.conversionSource,
-  }
-
-  const changeSummary = getExpenseChangeSummary(
-    toExpenseDomainShape(existingExpense),
-    resolvedExpense,
-    changeCtx,
-  )
-
-  // Union of old + new participant IDs so update emails reach everyone
-  // who was on the expense, including those removed by the change.
-  const affectedParticipantIds = [
-    ...getAffectedParticipantIds({
-      oldExpense: toExpenseDomainShape(existingExpense),
-      newExpense: expense,
-    }),
-  ]
-
-  const removedDocuments = existingExpense.documents.filter(
-    (existingDoc) => !documents.some((doc) => doc.id === existingDoc.id),
-  )
-  // S3 document deletions moved to post-transaction best-effort cleanup below
-
-  // Handle items: delete stale, create/update incoming
-  const incomingItems = expense.items ?? []
-  const existingItems = existingExpense.items ?? []
-
-  const isLeavingItemized =
-    existingExpense.splitMode === 'ITEMIZED' &&
-    expense.splitMode !== 'ITEMIZED' &&
-    existingItems.some((i) => i.paidFor.length > 0)
-
-  const incomingIds = new Set(
-    incomingItems.filter((i) => i.id).map((i) => i.id!),
-  )
-  const existingItemIds = new Set(existingItems.map((i) => i.id))
-  const itemsToDelete = existingItems.filter((i) => !incomingIds.has(i.id))
-
-  const existingSeries = existingExpense.recurringSeries
-  const detachRecurrence = existingSeries !== null && recurrence === null
-  const isDeleteRecurrence =
-    detachRecurrence && options?.scope === 'THIS_AND_FUTURE'
-  const isUpdateRecurrence = existingSeries !== null && recurrence !== null
-  const isCreateRecurrence = existingSeries === null && recurrence !== null
-
-  const expensePaidFor =
-    expense.splitMode === 'ITEMIZED'
-      ? computePaidForFromItems(
-          expense.items ?? [],
-          [...participantIds],
-          conversion.originalAmount ?? expenseAmount,
-          expense.itemizedRemainder,
-          expenseId,
-        ).paidFor
-      : expense.paidFor
-
-  const recurrenceTemplate = recurrence
-    ? buildRecurringTemplate({
-        expense: { ...expense, paidForOverride: expensePaidFor },
-        conversion,
-      })
-    : null
-
-  // Exchange rates are date-sensitive. Resolve them before entering the
-  // write transaction, then apply the resulting snapshots while the series
-  // row is locked. Newly materialized rows will use the updated series
-  // template after this transaction commits.
-  const materializedFutureRows =
-    isUpdateRecurrence && options?.scope === 'THIS_AND_FUTURE'
-      ? await prisma.expense.findMany({
-          where: {
-            recurringSeriesId: existingSeries!.id,
-            recurrenceSequence: {
-              gte: existingExpense.recurrenceSequence ?? 1,
-            },
-          },
-          orderBy: { recurrenceSequence: 'asc' },
-          select: { id: true, expenseDate: true, recurrenceSequence: true },
-        })
-      : []
-  const materializedConversions = new Map<
-    string,
-    Awaited<ReturnType<typeof resolveConversion>>
-  >()
-  if (recurrenceTemplate) {
-    const conversionInput =
-      recurrenceTemplate.conversionSource === 'CUSTOM'
-        ? {
-            type: 'custom' as const,
-            currency: recurrenceTemplate.originalCurrency ?? '',
-            rate: recurrenceTemplate.conversionRate ?? 1,
-          }
-        : recurrenceTemplate.conversionSource === 'EXCHANGE'
-          ? {
-              type: 'exchange' as const,
-              currency: recurrenceTemplate.originalCurrency ?? '',
-            }
-          : undefined
-    for (const row of materializedFutureRows) {
-      materializedConversions.set(
-        row.id,
-        await resolveConversion(
-          { amount: recurrenceTemplate.amount, conversion: conversionInput },
-          {
-            ledgerCurrency: group.ledger.currencyCode ?? null,
-            expenseDate: row.expenseDate,
-          },
-        ),
-      )
-    }
-  }
-
-  // Transaction: activity log + all DB writes are atomic
-  const { activity, createdExpense } = await prisma.$transaction(async (tx) => {
-    if (existingExpense.recurringSeriesId) {
-      await tx.$queryRaw`SELECT id FROM "RecurringExpenseSeries" WHERE id = ${existingExpense.recurringSeriesId} FOR UPDATE`
-      if (options?.scope === 'THIS_AND_FUTURE' && existingSeries !== null) {
-        const lockedSeries = await tx.recurringExpenseSeries.findUnique({
-          where: { id: existingSeries.id },
-          select: { version: true },
-        })
-        if (!lockedSeries || lockedSeries.version !== existingSeries.version) {
-          throw new Error(
-            'Recurring expense series changed while it was being updated; retry the update',
-          )
-        }
-      }
-    }
-    let act: Awaited<ReturnType<typeof logActivity>> | null = null
-
-    if (changeSummary) {
-      act = await logActivity(
-        groupId,
-        {
-          type: 'EXPENSE_UPDATED',
-          actor: { type: 'ACCOUNT', id: actor.accountId },
-          subject: { type: 'EXPENSE', id: expenseId },
-          data: buildExpenseActivityData({
-            summary: expense.title,
-            title: expense.title,
-            amount: expenseAmount,
-            currencyCode: conversion.originalCurrency,
-            date: expenseDateStr,
-            changedFields: changeSummary.changedFields,
-            changes: changeSummary.changes,
-            affectedParticipants: affectedParticipantIds,
-            originalAmount: conversion.originalAmount ?? undefined,
-            conversionRate: conversion.conversionRate ?? undefined,
-            conversionSource: conversion.conversionSource,
-            ledgerCurrencyCode: group.ledger.currencyCode ?? null,
-          }),
-        },
-        tx,
-      )
-    }
-
-    if (isLeavingItemized) {
-      await tx.expenseItemPaidFor.deleteMany({
-        where: { expenseItem: { expenseId } },
-      })
-    }
-
-    if (itemsToDelete.length > 0) {
-      await tx.expenseItem.deleteMany({
-        where: { id: { in: itemsToDelete.map((i) => i.id) } },
-      })
-    }
-
-    for (const item of incomingItems) {
-      if (item.id && existingItemIds.has(item.id)) {
-        await tx.expenseItem.update({
-          where: { id: item.id },
-          data: {
-            title: item.title,
-            unitPrice: item.unitPrice,
-            quantity: item.quantity,
-            amount: item.amount,
-            splitMode: item.splitMode,
-          },
-        })
-        if (!isLeavingItemized) {
-          await tx.expenseItemPaidFor.deleteMany({
-            where: { expenseItemId: item.id },
-          })
-          if (item.paidFor.length > 0) {
-            await tx.expenseItemPaidFor.createMany({
-              data: item.paidFor.map((pf) => ({
-                expenseItemId: item.id!,
-                ledgerParticipantId: pf.participant,
-                shares: pf.shares,
-              })),
-            })
-          }
-        }
-      } else {
-        const itemId = item.id ?? randomId()
-        await tx.expenseItem.create({
-          data: {
-            id: itemId,
-            expenseId,
-            title: item.title,
-            unitPrice: item.unitPrice,
-            quantity: item.quantity,
-            amount: item.amount,
-            splitMode: item.splitMode,
-            ...(!isLeavingItemized && item.paidFor.length > 0
-              ? {
-                  paidFor: {
-                    createMany: {
-                      data: item.paidFor.map((pf) => ({
-                        ledgerParticipantId: pf.participant,
-                        shares: pf.shares,
-                      })),
-                    },
-                  },
-                }
-              : {}),
-          },
-        })
-      }
-    }
-
-    // Only manage `ExpenseItemizedRemainder` rows for ITEMIZED expenses.
-    // The remainder is semantically meaningless otherwise, and we
-    // proactively delete any pre-existing rows so leftover artifacts from
-    // past (buggy) edits are cleaned up the next time the expense is
-    // updated.
-    if (expense.splitMode === 'ITEMIZED') {
-      await tx.expenseItemizedRemainder.deleteMany({
-        where: { expenseId },
-      })
-      if (expense.itemizedRemainder) {
-        await tx.expenseItemizedRemainder.create({
-          data: {
-            expenseId,
-            splitMode: expense.itemizedRemainder.splitMode,
-            paidFor: {
-              createMany: {
-                data: expense.itemizedRemainder.paidFor.map((pf) => ({
-                  ledgerParticipantId: pf.participant,
-                  shares: pf.shares,
-                })),
-              },
-            },
-          },
-        })
-      }
-    }
-
-    const updated = await tx.expense.update({
-      where: { id: expenseId },
-      data: {
-        expenseDate: preserveRecurringDates
-          ? existingExpense.expenseDate
-          : expense.expenseDate,
-        amount: expenseAmount,
-        originalAmount: conversion.originalAmount,
-        originalCurrency: conversion.originalCurrency,
-        conversionRate: conversion.conversionRate,
-        conversionSource: conversion.conversionSource,
-        title: expense.title,
-        categoryId: expense.category,
-        paidBySplitMode: expense.paidBySplitMode,
-        ...(expense.paidByList.length > 0
-          ? {
-              paidByList: {
-                create: expense.paidByList
-                  .filter(
-                    (p) =>
-                      !existingExpense.paidByList.some(
-                        (pb) => pb.ledgerParticipantId === p.participant,
-                      ),
-                  )
-                  .map((paidBy) => ({
-                    ledgerParticipantId: paidBy.participant,
-                    shares: paidBy.shares,
-                  })),
-                update: expense.paidByList.map((paidBy) => ({
-                  where: {
-                    expenseId_ledgerParticipantId: {
-                      expenseId,
-                      ledgerParticipantId: paidBy.participant,
-                    },
-                  },
-                  data: {
-                    shares: paidBy.shares,
-                  },
-                })),
-                deleteMany: existingExpense.paidByList
-                  .filter(
-                    (paidBy) =>
-                      !expense.paidByList.some(
-                        (p) => p.participant === paidBy.ledgerParticipantId,
-                      ),
-                  )
-                  .map(({ ledgerParticipantId, shares }) => ({
-                    ledgerParticipantId,
-                    shares,
-                  })),
-              },
-            }
-          : {}),
-        splitMode: expense.splitMode,
-        paidFor: {
-          create: expensePaidFor
-            .filter(
-              (p) =>
-                !existingExpense.paidFor.some(
-                  (pp) => pp.ledgerParticipantId === p.participant,
-                ),
-            )
-            .map((paidFor) => ({
-              ledgerParticipantId: paidFor.participant,
-              shares: paidFor.shares,
-            })),
-          update: expensePaidFor.map((paidFor) => ({
-            where: {
-              expenseId_ledgerParticipantId: {
-                expenseId,
-                ledgerParticipantId: paidFor.participant,
-              },
-            },
-            data: {
-              shares: paidFor.shares,
-            },
-          })),
-          deleteMany: existingExpense.paidFor.filter(
-            (paidFor) =>
-              !expensePaidFor.some(
-                (pf) => pf.participant === paidFor.ledgerParticipantId,
-              ),
-          ),
-        },
-        ...(detachRecurrence
-          ? { recurringSeries: { disconnect: true }, recurrenceSequence: null }
-          : {}),
-        isReimbursement: expense.isReimbursement,
-        documents: {
-          connectOrCreate: documents.map((doc) => ({
-            create: { ...doc, ledgerId: group.ledgerId },
-            where: { id: doc.id },
-          })),
-          deleteMany: existingExpense.documents
-            .filter(
-              (existingDoc) =>
-                !documents.some((doc) => doc.id === existingDoc.id),
-            )
-            .map((doc) => ({
-              id: doc.id,
-            })),
-        },
-        notes: expense.notes,
-      },
-    })
-
-    if (isDeleteRecurrence && existingSeries) {
-      await tx.recurringExpenseSeries.update({
-        where: { id: existingSeries.id },
-        data: { status: 'CANCELLED' },
-      })
-    } else if ((isCreateRecurrence || isUpdateRecurrence) && recurrence) {
-      const seriesId = existingSeries?.id ?? randomId()
-      const template = recurrenceTemplate!
-      if (!existingSeries) {
-        await createSeriesForExpense({
-          tx,
-          seriesId,
-          ledgerId: group.ledgerId,
-          creatorAccountId: actor.accountId,
-          anchorDate: expense.expenseDate,
-          config: recurrence,
-          template,
-          boss: queueBoss,
-        })
-        await tx.expense.update({
-          where: { id: expenseId },
-          data: { recurringSeriesId: seriesId, recurrenceSequence: 1 },
-        })
-      } else if (options?.scope === 'THIS_AND_FUTURE') {
-        const anchorSequence = existingExpense.recurrenceSequence ?? 1
-        // `occurrencesCreated` is authoritative and monotonic. Do not infer
-        // progress from surviving expense rows: occurrence-only deletion can
-        // leave sequence gaps, and this-and-future delete-only intentionally
-        // removes later rows while keeping the series active.
-        const seriesProgress = await tx.recurringExpenseSeries.findUnique({
-          where: { id: existingSeries.id },
-          select: { occurrencesCreated: true },
-        })
-        if (!seriesProgress) throw new Error('Recurring series not found')
-        const latest = await tx.expense.findFirst({
-          where: { recurringSeriesId: existingSeries.id },
-          orderBy: { expenseDate: 'desc' },
-          select: { expenseDate: true },
-        })
-        const maxSequence = Math.max(
-          anchorSequence,
-          seriesProgress.occurrencesCreated,
-        )
-        let nextOrdinal = maxSequence - anchorSequence + 2
-        let nextOccurrenceDate = calculateRecurrenceDate(
-          expense.expenseDate,
-          recurrence.frequency,
-          recurrence.interval,
-          nextOrdinal,
-        )
-        while (
-          latest?.expenseDate &&
-          nextOccurrenceDate <= latest.expenseDate
-        ) {
-          nextOrdinal += 1
-          nextOccurrenceDate = calculateRecurrenceDate(
-            expense.expenseDate,
-            recurrence.frequency,
-            recurrence.interval,
-            nextOrdinal,
-          )
-        }
-        const completed =
-          (recurrence.end.type === 'COUNT' &&
-            maxSequence >= recurrence.end.count) ||
-          (recurrence.end.type === 'DATE' &&
-            nextOccurrenceDate > recurrence.end.endDate)
-        await tx.recurringExpenseSeries.update({
-          where: { id: existingSeries.id },
-          data: {
-            frequency: recurrence.frequency,
-            interval: recurrence.interval,
-            endType: recurrence.end.type,
-            occurrenceLimit:
-              recurrence.end.type === 'COUNT' ? recurrence.end.count : null,
-            endDate:
-              recurrence.end.type === 'DATE' ? recurrence.end.endDate : null,
-            anchorDate: expense.expenseDate,
-            anchorSequence,
-            occurrencesCreated: maxSequence,
-            nextOccurrenceDate,
-            nextOccurrenceOrdinal: nextOrdinal,
-            status: completed ? 'COMPLETED' : 'ACTIVE',
-            version: { increment: 1 },
-            template,
-            catchUpBatch: null,
-          },
-        })
-        for (const row of materializedFutureRows) {
-          // The edited occurrence was already updated above (including its
-          // documents). Future rows receive the new template while retaining
-          // their own date, recurrence identity, and attachments.
-          if (row.id === expenseId) continue
-          await updateMaterializedOccurrence(
-            tx,
-            row,
-            template,
-            materializedConversions.get(row.id) ?? conversion,
-          )
-        }
-        if (!completed) {
-          await enqueueMaterialization(
-            tx,
-            {
-              seriesId: existingSeries.id,
-              sequence: maxSequence + 1,
-              occurrenceDate: nextOccurrenceDate,
-            },
-            queueBoss,
-          )
-        }
-      }
-    }
-    return { activity: act, createdExpense: updated }
-  })
-
-  // Best-effort S3 cleanup — errors are logged but not propagated so they
-  // don't corrupt the perceived mutation outcome.
-  for (const doc of removedDocuments) {
-    try {
-      await deleteS3Object(doc.url)
-    } catch (err) {
-      console.warn(`[expenses] failed to delete S3 object ${doc.url}:`, err)
-    }
-  }
-  if (activity && changeSummary) {
-    scheduleDefaultNotificationDispatch({
-      activityId: activity.id,
-      type: 'EXPENSE_UPDATED',
-      groupId,
-      actor: { type: 'ACCOUNT', id: actor.accountId },
-      subject: { type: 'EXPENSE', id: expenseId },
-      data: buildExpenseActivityData({
-        summary: expense.title,
-        title: expense.title,
-        amount: expenseAmount,
-        currencyCode: conversion.originalCurrency,
-        date: expenseDateStr,
-        changedFields: changeSummary.changedFields,
-        changes: changeSummary.changes,
-        affectedParticipants: affectedParticipantIds,
-        originalAmount: conversion.originalAmount ?? undefined,
-        conversionRate: conversion.conversionRate ?? undefined,
-        conversionSource: conversion.conversionSource,
-        ledgerCurrencyCode: group.ledger.currencyCode ?? null,
-      }),
-      occurredAt: activity.time,
-    })
-  }
-
-  return createdExpense
-}
-
-export async function getGroupExpensesParticipants(groupId: string) {
-  const expenses = await getGroupExpenses(groupId)
-  return Array.from(
-    new Set(
-      expenses.flatMap((e) => [
-        ...e.paidByList.map((pb) => pb.ledgerParticipant.id),
-        ...e.paidFor.map((pf) => pf.ledgerParticipant.id),
-      ]),
-    ),
-  )
-}
-
-type GetGroupExpensesSortBy = 'expenseDate' | 'createdAt' | 'amount'
-type GetGroupExpensesSortDir = 'asc' | 'desc'
-type GetGroupExpensesMatch = 'any' | 'all' | 'exact'
-
-type GetGroupExpensesOptions = {
-  offset?: number
-  length?: number
-  filter?: string
-  hideReimbursements?: boolean
-  categories?: string[]
-  paidBy?: string[]
-  paidByMatch?: GetGroupExpensesMatch
-  paidFor?: string[]
-  paidForMatch?: GetGroupExpensesMatch
-  dateFrom?: Date
-  dateTo?: Date
-  minAmount?: number
-  maxAmount?: number
-  currencies?: string[]
-  sortBy?: GetGroupExpensesSortBy
-  sortDir?: GetGroupExpensesSortDir
-}
-
-function buildParticipantMatch(
-  selected: string[] | undefined,
-  match: GetGroupExpensesMatch | undefined,
-  relation: 'paidByList' | 'paidFor',
-): Prisma.ExpenseWhereInput | undefined {
-  if (!selected || selected.length === 0) return undefined
-  const relationFilter = {
-    [relation]: {
-      some: { ledgerParticipantId: { in: selected } },
-    },
-  } as const
-  if ((match ?? 'any') === 'any') return relationFilter
-  if (match === 'exact') {
-    return {
-      AND: selected.map((id) => ({
-        [relation]: { some: { ledgerParticipantId: id } },
-      })),
-      NOT: {
-        [relation]: { some: { ledgerParticipantId: { notIn: selected } } },
-      },
-    } as Prisma.ExpenseWhereInput
-  }
-  return {
-    AND: selected.map((id) => ({
-      [relation]: { some: { ledgerParticipantId: id } },
-    })),
-  } as Prisma.ExpenseWhereInput
-}
-
-export async function getGroupExpenses(
-  groupId: string,
-  options?: GetGroupExpensesOptions,
-) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: { ledgerId: true },
-  })
-  if (!group?.ledgerId) return []
-
-  const expenseDateRange: Prisma.DateTimeFilter | undefined =
-    options?.dateFrom || options?.dateTo
-      ? {
-          ...(options.dateFrom ? { gte: options.dateFrom } : {}),
-          ...(options.dateTo ? { lte: options.dateTo } : {}),
-        }
-      : undefined
-
-  const amountRange: Prisma.IntFilter | undefined =
-    options?.minAmount !== undefined || options?.maxAmount !== undefined
-      ? {
-          ...(options.minAmount !== undefined
-            ? { gte: options.minAmount }
-            : {}),
-          ...(options.maxAmount !== undefined
-            ? { lte: options.maxAmount }
-            : {}),
-        }
-      : undefined
-
-  const where: Prisma.ExpenseWhereInput = {
-    ledgerId: group.ledgerId,
-    title: options?.filter
-      ? { contains: options.filter, mode: 'insensitive' }
-      : undefined,
-    isReimbursement: options?.hideReimbursements ? false : undefined,
-    categoryId:
-      options?.categories && options.categories.length > 0
-        ? { in: options.categories }
-        : undefined,
-    originalCurrency:
-      options?.currencies && options.currencies.length > 0
-        ? { in: options.currencies }
-        : undefined,
-    expenseDate: expenseDateRange,
-    amount: amountRange,
-    ...(buildParticipantMatch(
-      options?.paidBy,
-      options?.paidByMatch,
-      'paidByList',
-    ) ?? {}),
-    ...(buildParticipantMatch(
-      options?.paidFor,
-      options?.paidForMatch,
-      'paidFor',
-    ) ?? {}),
-  }
-
-  const sortField = options?.sortBy ?? 'expenseDate'
-  const sortDir = options?.sortDir ?? 'desc'
-  const primaryOrder: Prisma.ExpenseOrderByWithRelationInput = {
-    [sortField]: sortDir,
-  }
-  const orderBy: Prisma.ExpenseOrderByWithRelationInput[] =
-    sortField === 'expenseDate'
-      ? [primaryOrder, { createdAt: 'desc' }]
-      : [primaryOrder, { id: 'desc' }]
-
-  const rows = await prisma.expense.findMany({
-    select: {
-      amount: true,
-      conversionRate: true,
-      conversionSource: true,
-      categoryId: true,
-      createdAt: true,
-      expenseDate: true,
-      id: true,
-      recurrenceSequence: true,
-      isReimbursement: true,
-      originalAmount: true,
-      originalCurrency: true,
-      paidBySplitMode: true,
-      paidByList: {
-        select: {
-          ledgerParticipant: {
-            select: {
-              id: true,
-              groupMember: {
-                select: {
-                  account: { select: { id: true, name: true, image: true } },
-                },
-              },
-              invitations: {
-                select: { email: true, temporaryName: true },
-                take: 1,
-                orderBy: { createdAt: 'desc' },
-              },
-            },
-          },
-          shares: true,
-        },
-      },
-      paidFor: {
-        select: {
-          ledgerParticipant: {
-            select: {
-              id: true,
-              groupMember: {
-                select: {
-                  account: { select: { id: true, name: true, image: true } },
-                },
-              },
-              invitations: {
-                select: { email: true, temporaryName: true },
-                take: 1,
-                orderBy: { createdAt: 'desc' },
-              },
-            },
-          },
-          shares: true,
-        },
-      },
-      splitMode: true,
-      recurringSeries: {
-        select: {
-          id: true,
-          frequency: true,
-          interval: true,
-          endType: true,
-          occurrenceLimit: true,
-          endDate: true,
-          status: true,
-          anchorDate: true,
-          nextOccurrenceDate: true,
-        },
-      },
-      title: true,
-      _count: { select: { documents: true } },
-      items: {
-        select: {
-          id: true,
-          title: true,
-          unitPrice: true,
-          quantity: true,
-          amount: true,
-          splitMode: true,
-          paidFor: {
-            select: {
-              ledgerParticipantId: true,
-              shares: true,
-            },
-          },
-        },
-      },
-      itemizedRemainder: {
-        select: {
-          splitMode: true,
-          paidFor: {
-            select: {
-              ledgerParticipantId: true,
-              shares: true,
-            },
-          },
-        },
-      },
-    },
-    where,
-    orderBy,
-    skip: options && options.offset,
-    take: options && options.length,
-  })
-
-  return rows.map((row) => ({
-    ...row,
-    paidByList: row.paidByList.map((pb) => ({
-      ledgerParticipant: {
-        id: pb.ledgerParticipant.id,
-        name: resolveParticipantDisplayName(pb.ledgerParticipant),
-        account: pb.ledgerParticipant.groupMember?.account ?? null,
-      },
-      shares: pb.shares,
-    })),
-    paidFor: row.paidFor.map((pf) => ({
-      ledgerParticipant: {
-        id: pf.ledgerParticipant.id,
-        name: resolveParticipantDisplayName(pf.ledgerParticipant),
-        account: pf.ledgerParticipant.groupMember?.account ?? null,
-      },
-      shares: pf.shares,
-    })),
-    items: (row.items ?? []).map((item) => ({
-      id: item.id,
-      title: item.title,
-      unitPrice: item.unitPrice,
-      quantity: item.quantity,
-      amount: item.amount,
-      splitMode: item.splitMode,
-      paidFor: item.paidFor.map((pf) => ({
-        participant: pf.ledgerParticipantId,
-        shares: pf.shares,
-      })),
-    })),
-    itemizedRemainder: row.itemizedRemainder
-      ? {
-          splitMode: row.itemizedRemainder.splitMode,
-          paidFor: row.itemizedRemainder.paidFor.map((pf) => ({
-            participant: pf.ledgerParticipantId,
-            shares: pf.shares,
-          })),
-        }
-      : undefined,
-    categoryId: narrowCategoryId(row.categoryId),
-    category: resolveCategory(row.categoryId),
-    conversionRate: row.conversionRate ?? null,
-    conversionSource: row.conversionSource,
-    recurringSeriesId: row.recurringSeries?.id ?? null,
-    recurrenceSequence: row.recurrenceSequence,
-    recurrence: row.recurringSeries
-      ? toRecurrenceConfig(row.recurringSeries)
-      : null,
-    recurringSeriesStatus: row.recurringSeries?.status ?? null,
-  }))
-}
-
-/** Replace the mutable template fields on a materialized occurrence. Dates,
- * recurrence identity, and documents intentionally remain untouched. */
-async function updateMaterializedOccurrence(
-  tx: Prisma.TransactionClient,
-  row: { id: string },
-  template: ReturnType<typeof buildRecurringTemplate>,
-  conversion: {
-    ledgerAmountMinor: number
-    originalAmount: number | null
-    originalCurrency: string | null
-    conversionRate: number | null
-    conversionSource: 'EXCHANGE' | 'CUSTOM' | null
-  },
-) {
-  await tx.expenseItemizedRemainder.deleteMany({ where: { expenseId: row.id } })
-  await tx.expense.update({
-    where: { id: row.id },
-    data: {
-      amount: conversion.ledgerAmountMinor,
-      originalAmount: conversion.originalAmount,
-      originalCurrency: conversion.originalCurrency,
-      conversionRate: conversion.conversionRate,
-      conversionSource: conversion.conversionSource,
-      title: template.title,
-      categoryId: template.categoryId,
-      paidBySplitMode: template.paidBySplitMode as never,
-      splitMode: template.splitMode as never,
-      isReimbursement: template.isReimbursement,
-      notes: template.notes,
-      paidByList: {
-        deleteMany: {},
-        createMany: { data: template.paidByList },
-      },
-      paidFor: {
-        deleteMany: {},
-        createMany: { data: template.paidFor },
-      },
-      items: {
-        deleteMany: {},
-        create: template.items.map((item) => ({
-          id: randomId(),
-          title: item.title,
-          unitPrice: item.unitPrice,
-          quantity: item.quantity,
-          amount: item.amount,
-          splitMode: item.splitMode as never,
-          paidFor: { createMany: { data: item.paidFor } },
-        })),
-      },
-    },
-  })
-  if (template.itemizedRemainder) {
-    await tx.expenseItemizedRemainder.create({
-      data: {
-        expenseId: row.id,
-        splitMode: template.itemizedRemainder.splitMode as never,
-        paidFor: {
-          createMany: { data: template.itemizedRemainder.paidFor },
-        },
-      },
-    })
-  }
-}
-
-export async function getGroupExpenseCount(groupId: string) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: { ledgerId: true },
-  })
-  if (!group?.ledgerId) return 0
-  return prisma.expense.count({ where: { ledgerId: group.ledgerId } })
-}
-
-/**
- * Rank currencies previously used in the group (excluding the pinned group
- * ledger currency). Loads only `originalCurrency` + `expenseDate` within a
- * recency lookback so large ledgers stay cheap; scoring uses a 90-day
- * half-life (see `@spliit/domain` `rankCommonCurrencies`).
- */
-export async function getGroupCommonCurrencies(groupId: string) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: {
-      ledgerId: true,
-      ledger: { select: { currencyCode: true } },
-    },
-  })
-  if (!group?.ledgerId) return [] as string[]
-
-  const since = commonCurrencyLookbackDate()
-  const rows = await prisma.expense.findMany({
-    where: {
-      ledgerId: group.ledgerId,
-      expenseDate: { gte: since },
-    },
-    select: {
-      originalCurrency: true,
-      expenseDate: true,
-    },
-  })
-
-  return rankCommonCurrencies(rows, {
-    groupCurrency: group.ledger.currencyCode,
-  })
-}
-
-export async function getExpense(groupId: string, expenseId: string) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: { ledgerId: true },
-  })
-  if (!group?.ledgerId) return null
-  const expense = await prisma.expense.findFirst({
-    where: { id: expenseId, ledgerId: group.ledgerId },
-    include: {
-      paidByList: { include: { ledgerParticipant: true } },
-      paidFor: true,
-      documents: true,
-      recurringSeries: true,
-      items: {
-        include: { paidFor: true },
-      },
-      itemizedRemainder: {
-        include: { paidFor: true },
-      },
-    },
-  })
-  if (!expense) return null
-  const previousExpense =
-    expense.recurringSeries && expense.recurrenceSequence
-      ? await prisma.expense.findFirst({
-          where: {
-            recurringSeriesId: expense.recurringSeriesId,
-            recurrenceSequence: { lt: expense.recurrenceSequence },
-          },
-          orderBy: { recurrenceSequence: 'desc' },
-          select: { id: true },
-        })
-      : null
-  const nextExpense =
-    expense.recurringSeries && expense.recurrenceSequence
-      ? await prisma.expense.findFirst({
-          where: {
-            recurringSeriesId: expense.recurringSeriesId,
-            recurrenceSequence: { gt: expense.recurrenceSequence },
-          },
-          orderBy: { recurrenceSequence: 'asc' },
-          select: { id: true },
-        })
-      : null
-  return {
-    ...expense,
-    categoryId: narrowCategoryId(expense.categoryId),
-    category: resolveCategory(expense.categoryId),
-    recurrence: expense.recurringSeries
-      ? toRecurrenceConfig(expense.recurringSeries)
-      : null,
-    previousExpenseId: previousExpense?.id ?? null,
-    nextExpenseId: nextExpense?.id ?? null,
-  }
-}
-
-export async function getRecurringExpenseSeries(
-  groupId: string,
-  options?: {
-    cursor?: string
-    limit?: number
-    seriesId?: string
-    occurrenceCursor?: number
-    occurrenceLimit?: number
-  },
-) {
-  const group = await prisma.group.findUnique({
-    where: { id: groupId },
-    select: { ledgerId: true },
-  })
-  if (!group?.ledgerId) return { series: [], nextCursor: null }
-  const limit = Math.min(Math.max(options?.limit ?? 20, 1), 100)
-  const rows = await prisma.recurringExpenseSeries.findMany({
-    where: {
-      ledgerId: group.ledgerId,
-      ...(options?.seriesId ? { id: options.seriesId } : {}),
-    },
-    orderBy: [{ anchorDate: 'desc' }, { id: 'desc' }],
-    ...(options?.cursor ? { cursor: { id: options.cursor }, skip: 1 } : {}),
-    take: limit + 1,
-    include: {
-      expenses: options?.seriesId
-        ? {
-            where: options.occurrenceCursor
-              ? { recurrenceSequence: { gt: options.occurrenceCursor } }
-              : undefined,
-            orderBy: { recurrenceSequence: 'asc' },
-            take: Math.min(Math.max(options.occurrenceLimit ?? 50, 1), 100) + 1,
-            select: {
-              id: true,
-              expenseDate: true,
-              recurrenceSequence: true,
-              title: true,
-              amount: true,
-            },
-          }
-        : false,
-    },
-  })
-  const hasMore = rows.length > limit
-  const page = rows.slice(0, limit)
-  return {
-    series: page.map((series) => {
-      const expenses = Array.isArray(series.expenses) ? series.expenses : []
-      const occurrenceLimit = options?.occurrenceLimit ?? 50
-      return {
-        id: series.id,
-        frequency: series.frequency,
-        interval: series.interval,
-        anchorDate: series.anchorDate,
-        nextOccurrenceDate: series.nextOccurrenceDate,
-        endType: series.endType,
-        occurrenceLimit: series.occurrenceLimit,
-        endDate: series.endDate,
-        occurrencesCreated: series.occurrencesCreated,
-        status: series.status,
-        recurrence: toRecurrenceConfig(series),
-        expenses: expenses.slice(0, occurrenceLimit),
-        hasMoreOccurrences: expenses.length > occurrenceLimit,
-        nextOccurrenceCursor:
-          expenses.length > occurrenceLimit
-            ? (expenses[occurrenceLimit - 1]?.recurrenceSequence ?? null)
-            : null,
-      }
-    }),
-    nextCursor: hasMore ? (page.at(-1)?.id ?? null) : null,
-  }
-}
