+++ apps/api/src/lib/api/expenses/index.ts
@@ -1,0 +1,12 @@
+export { createExpense } from './create-expense'
+export { deleteExpense } from './delete-expense'
+export {
+  getExpense,
+  getGroupCommonCurrencies,
+  getGroupExpenseCount,
+  getGroupExpenses,
+  getGroupExpensesParticipants,
+  getRecurringExpenseSeries,
+} from './queries'
+export { stopRecurrence } from './stop-recurrence'
+export { updateExpense } from './update-expense'
