+++ apps/api/src/lib/api/index.ts
@@ -60,4 +60,8 @@ export {
   removeMember,
   updateMemberRole,
 } from './members'
+export {
+  getRecurringSeriesProgress,
+  type RecurringSeriesProgress,
+} from './series-progress'
 export { randomId, type GroupWithLedger } from './shared'
